Move Option and ParameterData to seperate file
